+++ allink_core/core/views.py
@@ -146,7 +146,11 @@ class AllinkBaseDetailView(TranslatableSlugMixin, DetailView):
             context.update({'base_template': 'app_content/ajax_base.html'})
         else:
             context.update({'base_template': 'base.html'})
-        return render_to_response(self.get_template_names(), context, context_instance=RequestContext(self.request))
+
+        categories = context['object'].categories.all()
+        app_label = context['object']._meta.app_label
+        templates = ['{}/{}_{}_detail.html'.format(app_label, app_label, category.slug) for category in categories] + self.get_template_names()
+        return render_to_response(templates, context, context_instance=RequestContext(self.request))
 
 
 class AllinkBaseCreateView(CreateView):